Improve signature of methods to work with template.
Description: I implemented the idea from the previous commit (added String() to filters), resulting in a simplified apply() signature. Also removed unused Template fields.

Idea: Make a sorting structure into which to take the sorting order and criterion, create a String method that returns a string with a criterion and order, thus delegating the sorting part of logic from the templates.