Update the aggregator logic.
Description: When implementing the server part of the news aggregator, I encountered a problem with parsing resources. Before the changes I file as the corresponding source request. After the changes, I initialize all the resources at the program start and save them to cache.

Problems: Documentation for all modified components needs to be updated.